This program is an assignment for the first week of the SCA Mentorship Cohort 4
A program where the computer randomly generates a number between 0 and 20. The user needs to guess what the number is. If the user guesses wrong, tell them their guess is either too high, or too low.
This program is an assignment for the first week of the SCA Mentorship Programme. Cohort 4